Revert 4b56102..9a3681e
This rolls back to commit 4b561020f175f8b2eff4d1f61a62ebd47aee6bb7.

+++ src/registros/registros.h
@@ -8,7 +8,7 @@
 typedef struct 
 {
 	char nombre[4];			
-	float data;
+	int data;
 } registro;
 
 extern registro registros[14];
